==> project-2/Notebooks/01_wf_extract_load_source_raw.ipynb <==
1 - IMPORT MAIN LIBRARIES AND CREATE GLOBAL VARIABLES

import pandas as pd
import os
import requests

source_path = '..\\Source\\'
raw_path = '..\\Raw\\'
api_url = 'https://olinda.bcb.gov.br/olinda/servico/Informes_ListaTarifasPorInstituicaoFinanceira/versao/v1/odata/ListaTarifasPorInstituicaoFinanceira(PessoaFisicaOuJuridica=@PessoaFisicaOuJuridica,CNPJ=@CNPJ)?%40PessoaFisicaOuJuridica=%27J%27&%40CNPJ=%27REPLACECNPJ%27'

2 - EXTRACT AND LOAD CSV DATA - RAW LAYER

# Get complete path for all CSV files in "source" directory
filepaths = [source_path + file for file in os.listdir(source_path) if file.endswith('.csv')]

# Concat multiple CSV files in one data frame
df_raw_banks = pd.concat([pd.read_csv(filepath, delimiter=';', encoding='cp1252') for filepath in filepaths])
df_raw_banks.info()

# Load combined CSV to raw folder
df_raw_banks.to_csv(raw_path + 'banks_raw_data.csv', sep=',', encoding='cp1252')

3 - EXTRACT AND LOAD API DATA - RAW LAYER

# Execute API for each bank extracted
responseList = []
bankList = df_raw_banks[df_raw_banks['CNPJ IF'].str.strip() != '']

for cnpj in bankList['CNPJ IF']:
    response = requests.get(api_url.replace('REPLACECNPJ', cnpj))
    json = response.json()
    df_response = pd.DataFrame(json['value'])
    df_response['CNPJ'] = cnpj
    responseList.append(df_response)

df_raw_taxes = pd.concat(bank_taxes for bank_taxes in responseList)
df_raw_taxes.to_csv(raw_path + 'taxes_raw_data.csv', sep=',', encoding='cp1252')
df_raw_taxes.info()
