==> project-2/Notebooks/02_wf_transform_raw_trusted.ipynb <==
1 - IMPORT MAIN LIBRARIES AND CREATE GLOBAL VARIABLES

import pandas as pd
from datetime import datetime
import unicodedata

raw_path = '..\\Raw\\'
trusted_path = '..\\Trusted\\'

2.0 - Transform Bank Data -> Dataset Counts

# Get banks raw CSV data
df_trusted_banks = pd.read_csv(raw_path + 'banks_raw_data.csv', encoding='cp1252')
df_trusted_banks.count()

2.1 - Transform Bank Data -> Fix Columns

# Drop blank column and rename unnamed column
df_trusted_banks.drop(df_trusted_banks.columns[15], axis=1, inplace=True)
df_trusted_banks.rename(columns={'Unnamed: 0': 'id'}, inplace=True)
df_trusted_banks.keys()

2.2 - Transform Bank Data -> Data Types

# Remove blank strings and convert numeric values
df_trusted_banks = df_trusted_banks.astype(str)
df_trusted_banks["Índice"] = df_trusted_banks["Índice"].str.strip().replace('', None)
df_trusted_banks["Quantidade de reclamações reguladas procedentes"] = df_trusted_banks["Quantidade de reclamações reguladas procedentes"].str.strip().replace('', None)
df_trusted_banks["Quantidade de reclamações reguladas - outras"] = df_trusted_banks["Quantidade de reclamações reguladas - outras"].str.strip().replace('', None)
df_trusted_banks["Quantidade de reclamações não reguladas"] = df_trusted_banks["Quantidade de reclamações não reguladas"].str.strip().replace('', None)
df_trusted_banks["Quantidade total de reclamações"] = df_trusted_banks["Quantidade total de reclamações"].str.strip().replace('', None)
df_trusted_banks["Quantidade total de clientes – CCS e SCR"] = df_trusted_banks["Quantidade total de clientes – CCS e SCR"].str.strip().replace('', None)
df_trusted_banks["Quantidade de clientes – CCS"] = df_trusted_banks["Quantidade de clientes – CCS"].str.strip().replace('', None)
df_trusted_banks["Quantidade de clientes – SCR"] = df_trusted_banks["Quantidade de clientes – SCR"].str.strip().replace('', None)

df_trusted_banks = df_trusted_banks.astype({
    'id': int,
    'Ano': int,
    'Trimestre': str,
    'Categoria': str,
    'Tipo': str,
    'CNPJ IF': str,
    'Instituição financeira': str,
    'Índice': str,
    'Quantidade de reclamações reguladas procedentes': int,
    'Quantidade de reclamações reguladas - outras': int,
    'Quantidade de reclamações não reguladas': int,
    'Quantidade total de reclamações': int,
    'Quantidade total de clientes – CCS e SCR': int,
    'Quantidade de clientes – CCS': int, 
    'Quantidade de clientes – SCR': int
})

df_trusted_banks.info()     

2.3 - Transform Bank Data -> Load to trusted

# Load data to trusted path
df_trusted_banks.to_csv(trusted_path + 'banks.csv', encoding='cp1252')

3.0 - Transform Taxes Data -> Dataset Counts

# Get taxes raw CSV data
df_trusted_taxes = pd.read_csv(raw_path + 'taxes_raw_data.csv', encoding='cp1252')
df_trusted_taxes.count()

3.1 - Transform Taxes Data -> Fix Columns

# Rename unnamed column 
df_trusted_taxes.rename(columns={'Unnamed: 0': 'id'}, inplace=True)
df_trusted_taxes.keys()

3.2 - Transform Taxes Data -> Data Types

# Fix encoding and data ypes
df_trusted_taxes = df_trusted_taxes.astype(str)
df_trusted_taxes['Unidade'] = df_trusted_taxes['Unidade'].str.normalize('NFKD').str.encode('ascii','ignore').str.decode('ascii','ignore').str.lower()

df_trusted_taxes = df_trusted_taxes.astype({
    'id': int,
    'CodigoServico': int,
    'Servico': str,
    'Unidade': str,
    'DataVigencia': str,
    'ValorMaximo': float,
    'TipoValor': str,
    'Periodicidade': str,
    'CNPJ': str
})
df_trusted_taxes.info()

3.3 - Transform Taxes Data -> Load to trusted

# Load data to trusted path
df_trusted_taxes.to_csv(trusted_path + 'taxes.csv', encoding='cp1252')
